pyproject.toml
[tool.poetry]
name = "quizapp"
version = "0.1.0"
description = ""
authors = ["Kapitoshka-a <taraspanechko737@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pytest = "^8.2.2"
black = "^24.4.2"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.31"}
alembic = "^1.13.2"
fastapi-cache2 = "^0.2.1"
redis = "^5.0.7"
pydantic-settings = "^2.3.4"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
